extracted DB and homeserver core from mod
